moved the traffic light class to params.py
