+++ flow/benchmarks/baselines/bottleneck2.py
@@ -8,7 +8,7 @@ from flow.core.experiment import SumoExperiment
 from flow.core.params import InitialConfig
 from flow.core.params import InFlows
 from flow.core.vehicles import Vehicles
-from flow.core.traffic_lights import TrafficLights
+from flow.core.params import TrafficLights
 from flow.controllers import ContinuousRouter
 from flow.benchmarks.bottleneck2 import flow_params
 from flow.benchmarks.bottleneck2 import SCALING
