+++ flow/benchmarks/baselines/figureeight012.py
@@ -7,7 +7,7 @@ import numpy as np
 from flow.core.experiment import SumoExperiment
 from flow.core.params import InitialConfig
 from flow.core.vehicles import Vehicles
-from flow.core.traffic_lights import TrafficLights
+from flow.core.params import TrafficLights
 from flow.controllers import IDMController
 from flow.controllers import ContinuousRouter
 from flow.benchmarks.figureeight0 import flow_params
