+++ flow/benchmarks/bottleneck2.py
@@ -13,7 +13,7 @@ Horizon: 1000 steps
 
 from flow.core.params import SumoParams, EnvParams, InitialConfig, NetParams, \
     InFlows
-from flow.core.traffic_lights import TrafficLights
+from flow.core.params import TrafficLights
 from flow.core.vehicles import Vehicles
 from flow.controllers import RLController, ContinuousRouter
