+++ flow/benchmarks/rllab/trpo_runner.py
@@ -13,7 +13,7 @@ from rllab.misc.instrument import run_experiment_lite
 from rllab.policies.gaussian_mlp_policy import GaussianMLPPolicy
 
 from flow.core.params import InitialConfig
-from flow.core.traffic_lights import TrafficLights
+from flow.core.params import TrafficLights
 
 # use this to specify the environment to run
 from flow.benchmarks.grid1 import flow_params
