+++ flow/scenarios/grid.py
@@ -2,7 +2,7 @@
 
 from flow.scenarios.base_scenario import Scenario
 from flow.core.params import InitialConfig
-from flow.core.traffic_lights import TrafficLights
+from flow.core.params import TrafficLights
 from collections import defaultdict
 
 ADDITIONAL_NET_PARAMS = {
