+++ flow/scenarios/loop.py
@@ -2,7 +2,7 @@
 
 from flow.scenarios.base_scenario import Scenario
 from flow.core.params import InitialConfig
-from flow.core.traffic_lights import TrafficLights
+from flow.core.params import TrafficLights
 from numpy import pi, sin, cos, linspace
 
 ADDITIONAL_NET_PARAMS = {
