+++ flow/scenarios/netfile.py
@@ -1,7 +1,7 @@
 """Contains the scenario class for .net.xml files."""
 
 from flow.core.params import InitialConfig
-from flow.core.traffic_lights import TrafficLights
+from flow.core.params import TrafficLights
 from flow.scenarios.base_scenario import Scenario
